+++ packages/azure-http-specs/specs/azure/core/page/mockapi.ts
@@ -1,4 +1,4 @@
-import { json, passOnSuccess, ScenarioMockApi } from "@typespec/spec-api";
+import { dyn, dynItem, json, passOnSuccess, ScenarioMockApi } from "@typespec/spec-api";
 
 export const Scenarios: Record<string, ScenarioMockApi> = {};
 const validUser = { id: 1, name: "Madge", etag: "11bdc430-65e8-45ad-81d9-8ffa60d55b59" };
@@ -63,9 +63,8 @@ Scenarios.Azure_Core_Page_withParameterizedNextLink = passOnSuccess([
     response: {
       status: 200,
       body: json({
-        value: [{ id: 1, name: "User1" }],
-        // Include the original query parameters in the nextLink
-        nextLink: "/azure/core/page/with-parameterized-next-link/second-page?select=name",
+        values: [{ id: 1, name: "User1" }],
+        nextLink: dyn`${dynItem("baseUrl")}/azure/core/page/with-parameterized-next-link/second-page?select=name`,
       }),
     },
     kind: "MockApiDefinition",
@@ -83,7 +82,7 @@ Scenarios.Azure_Core_Page_withParameterizedNextLink = passOnSuccess([
     response: {
       status: 200,
       body: json({
-        value: [{ id: 2, name: "User2" }],
+        values: [{ id: 2, name: "User2" }],
       }),
     },
     kind: "MockApiDefinition",