+++ packages/typespec-client-generator-core/src/http.ts
@@ -6,13 +6,11 @@ import {
   Type,
   compilerAssert,
   createDiagnosticCollector,
-  getDoc,
   getEncode,
   getSummary,
-  ignoreDiagnostics,
   isErrorModel,
 } from "@typespec/compiler";
-import { $ } from "@typespec/compiler/experimental/typekit";
+import { $ } from "@typespec/compiler/typekit";
 import {
   HttpOperation,
   HttpOperationParameter,
@@ -54,8 +52,9 @@ import {
   TCGCContext,
 } from "./interfaces.js";
 import {
-  findRootSourceProperty,
+  compareRootSourceProperties,
   getAvailableApiVersions,
+  getClientDoc,
   getHttpBodySpreadModel,
   getHttpOperationResponseHeaders,
   getLocationOfOperation,
@@ -89,6 +88,7 @@ export function getSdkHttpOperation(
   httpOperation: HttpOperation,
   methodParameters: SdkMethodParameter[],
 ): [SdkHttpOperation, readonly Diagnostic[]] {
+  const tk = $(context.program);
   const diagnostics = createDiagnosticCollector();
   const { responses, exceptions } = diagnostics.pipe(
     getSdkHttpResponseAndExceptions(context, httpOperation),
@@ -97,11 +97,11 @@ export function getSdkHttpOperation(
     // we make sure valid responses and 404 responses are booleans
     for (const response of responses) {
       // all valid responses will return boolean
-      response.type = getSdkConstant(context, $.literal.createBoolean(true));
+      response.type = getSdkConstant(context, tk.literal.createBoolean(true));
     }
     const fourOFourResponse = exceptions.find((e) => e.statusCodes === 404);
     if (fourOFourResponse) {
-      fourOFourResponse.type = getSdkConstant(context, $.literal.createBoolean(false));
+      fourOFourResponse.type = getSdkConstant(context, tk.literal.createBoolean(false));
       // move from exception to valid response with status code 404
       responses.push({
         ...fourOFourResponse,
@@ -114,7 +114,7 @@ export function getSdkHttpOperation(
       responses.push({
         kind: "http",
         statusCodes: 404,
-        type: getSdkConstant(context, $.literal.createBoolean(false)),
+        type: getSdkConstant(context, tk.literal.createBoolean(false)),
         apiVersions: getAvailableApiVersions(
           context,
           httpOperation.operation,
@@ -224,12 +224,6 @@ function getSdkHttpParameters(
             httpOperation.operation,
           ),
         );
-        optional =
-          tspBody.type.kind === "Model"
-            ? [...tspBody.type.properties.values()].some((p) => !p.optional)
-              ? false
-              : true
-            : false;
       } else {
         type = diagnostics.pipe(
           getClientTypeWithDiagnostics(context, tspBody.type, httpOperation.operation),
@@ -242,7 +236,7 @@ function getSdkHttpParameters(
         name,
         isGeneratedName: true,
         serializedName: "",
-        doc: getDoc(context.program, tspBody.type),
+        doc: getClientDoc(context, tspBody.type),
         summary: getSummary(context.program, tspBody.type),
         onClient: false,
         contentTypes: [],
@@ -434,13 +428,7 @@ export function getSdkHttpParameter(
       // url type need allow reserved
       allowReserved:
         (httpParam as HttpOperationPathParameter)?.allowReserved ??
-        ignoreDiagnostics(
-          program.checker.isTypeAssignableTo(
-            param.type,
-            program.checker.getStdType("url"),
-            param.type,
-          ),
-        ),
+        $(program).type.isAssignableTo(param.type, $(program).builtin.url, param.type),
       serializedName: getPathParamName(program, param) ?? base.name,
       correspondingMethodParams,
       optional: param.optional,
@@ -723,7 +711,7 @@ function findMapping(
     if (
       methodParam.__raw &&
       serviceParam.__raw &&
-      findRootSourceProperty(methodParam.__raw) === findRootSourceProperty(serviceParam.__raw)
+      compareRootSourceProperties(methodParam.__raw, serviceParam.__raw)
     ) {
       return methodParam;
     }
@@ -812,7 +800,7 @@ function getFormatFromExplodeOrEncode(
   explode?: boolean,
 ): [CollectionFormat | undefined, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
-  if ($.array.is(type.type)) {
+  if ($(context.program).array.is(type.type)) {
     if (explode) {
       return diagnostics.wrap("multi");
     }