+++ packages/typespec-client-generator-core/test/clients/structure.test.ts
@@ -92,7 +92,7 @@ it("arm client with operation groups", async () => {
 
   const tests = client.children?.find((c) => c.name === "Tests");
   ok(tests);
-  strictEqual(tests.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(tests.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(tests.clientInitialization.parameters.length, 4);
   strictEqual(tests.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(tests.clientInitialization.parameters[1].name, "credential");
@@ -150,7 +150,7 @@ it("client with sub clients", async () => {
 
   const pets = client.children?.find((c) => c.name === "Pets");
   ok(pets);
-  strictEqual(pets.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(pets.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(pets.clientInitialization.parameters.length, 1);
   strictEqual(pets.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(pets?.methods.length, 0);
@@ -158,7 +158,7 @@ it("client with sub clients", async () => {
 
   const dogs = pets.children?.find((c) => c.name === "Dogs");
   ok(dogs);
-  strictEqual(dogs.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(dogs.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(dogs.clientInitialization.parameters.length, 1);
   strictEqual(dogs.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(dogs?.methods.length, 2);
@@ -167,7 +167,7 @@ it("client with sub clients", async () => {
 
   const cats = pets.children?.find((c) => c.name === "Cats");
   ok(cats);
-  strictEqual(cats.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(cats.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(cats.clientInitialization.parameters.length, 1);
   strictEqual(cats.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(cats?.methods.length, 2);
@@ -176,7 +176,7 @@ it("client with sub clients", async () => {
 
   const actions = client.children?.find((c) => c.name === "Actions");
   ok(actions);
-  strictEqual(actions.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(actions.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(actions.clientInitialization.parameters.length, 1);
   strictEqual(actions.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(actions?.methods.length, 2); // client accessor methods which have already deprecated
@@ -244,7 +244,7 @@ it("client with sub client and sub client has extra initialization paramters", a
 
   const largeFaceList = client.children?.find((c) => c.name === "LargeFaceList");
   ok(largeFaceList);
-  strictEqual(largeFaceList.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(largeFaceList.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(largeFaceList.clientInitialization.parameters.length, 2);
   strictEqual(largeFaceList.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(largeFaceList.clientInitialization.parameters[1].name, "largeFaceListId");
@@ -254,7 +254,7 @@ it("client with sub client and sub client has extra initialization paramters", a
 
   const largePersonGroup = client.children?.find((c) => c.name === "LargePersonGroup");
   ok(largePersonGroup);
-  strictEqual(largePersonGroup.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(largePersonGroup.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(largePersonGroup.clientInitialization.parameters.length, 2);
   strictEqual(largePersonGroup.clientInitialization.parameters[0].name, "endpoint");
   strictEqual(largePersonGroup.clientInitialization.parameters[1].name, "largePersonGroupId");
@@ -424,6 +424,7 @@ it("sub client could not only be initialized individually", async () => {
       "Invalid 'initializedBy' value. Sub client must have `InitializedBy.parent` or `InitializedBy.individually | InitializedBy.parent` specified in `initializedBy`.",
   });
 });
+
 it("single with core", async () => {
   const runnerWithCore = await createSdkTestRunner({
     librariesToAdd: [AzureCoreTestLibrary],
@@ -580,12 +581,68 @@ it("namespace", async () => {
     @server("http://localhost:3000", "endpoint")
     @useAuth(ApiKeyAuth<ApiKeyLocation.header, "x-ms-api-key">)
     @service
-    namespace My.Service {};
+    namespace My.Service;
+    op func(): void;
   `);
   const sdkPackage = runnerWithCore.context.sdkPackage;
   strictEqual(sdkPackage.clients.length, 1);
   const clientOne = sdkPackage.clients.filter((c) => c.name === "ServiceClient")[0];
   strictEqual(clientOne.namespace, "My.Service");
+});
+
+it("model-only namespace should be filtered out", async () => {
+  const runnerWithCore = await createSdkTestRunner({
+    librariesToAdd: [AzureCoreTestLibrary],
+    autoUsings: ["Azure.Core"],
+    emitterName: "@azure-tools/typespec-java",
+  });
+  await runnerWithCore.compile(`
+    @service
+    namespace Foo {
+      @usage(Usage.input)
+      model B {}
+    }
+  `);
+  const sdkPackage = runnerWithCore.context.sdkPackage;
+  strictEqual(sdkPackage.clients.length, 0);
+  strictEqual(sdkPackage.models.length, 1);
+});
+
+it("empty namespace with empty subclient", async () => {
+  const runnerWithCore = await createSdkTestRunner({
+    librariesToAdd: [AzureCoreTestLibrary],
+    autoUsings: ["Azure.Core"],
+    emitterName: "@azure-tools/typespec-java",
+  });
+  await runnerWithCore.compile(`
+    @service
+    namespace Foo {
+      model B {}
+      namespace Bar {
+        model A {}
+      }
+      interface Baz {}
+    }
+  `);
+  const sdkPackage = runnerWithCore.context.sdkPackage;
+  strictEqual(sdkPackage.clients.length, 0);
+});
+
+it("explicit clients with only models should not be filtered out", async () => {
+  const runnerWithCore = await createSdkTestRunner({
+    librariesToAdd: [AzureCoreTestLibrary],
+    autoUsings: ["Azure.Core"],
+    emitterName: "@azure-tools/typespec-java",
+  });
+  await runnerWithCore.compile(`
+    @client
+    @service
+    namespace Foo {
+      model B {}
+    }
+  `);
+  const sdkPackage = runnerWithCore.context.sdkPackage;
+  strictEqual(sdkPackage.clients.length, 1);
 });
 
 it("operationGroup", async () => {
@@ -612,7 +669,7 @@ it("operationGroup", async () => {
   strictEqual(mainClient.crossLanguageDefinitionId, "TestService");
 
   strictEqual(operationGroup.clientInitialization.parameters.length, 1);
-  strictEqual(operationGroup.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(operationGroup.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(operationGroup.methods.length, 1);
   strictEqual(operationGroup.methods[0].name, "func");
   strictEqual(
@@ -657,13 +714,13 @@ it("operationGroup2", async () => {
   strictEqual(mainClient.crossLanguageDefinitionId, "TestService");
 
   strictEqual(fooClient.clientInitialization.parameters.length, 1);
-  strictEqual(fooClient.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(fooClient.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(fooClient.methods.length, 0);
   strictEqual(fooClient.children?.length, 1);
   strictEqual(fooClient.crossLanguageDefinitionId, "TestService.Foo");
 
   strictEqual(fooBarClient.clientInitialization.parameters.length, 1);
-  strictEqual(fooBarClient.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(fooBarClient.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(fooBarClient.crossLanguageDefinitionId, "TestService.Foo.Bar");
   strictEqual(fooBarClient.methods.length, 1);
   strictEqual(fooBarClient.methods[0].kind, "basic");
@@ -671,7 +728,7 @@ it("operationGroup2", async () => {
   strictEqual(fooBarClient.methods[0].crossLanguageDefinitionId, "TestService.Foo.Bar.one");
 
   strictEqual(barClient.clientInitialization.parameters.length, 1);
-  strictEqual(barClient.clientInitialization.initializedBy, InitializedByFlags.Parent);
+  strictEqual(barClient.clientInitialization.initializedBy, InitializedByFlags.Default);
   strictEqual(barClient.crossLanguageDefinitionId, "TestService.Bar");
   strictEqual(barClient.methods.length, 1);
   strictEqual(barClient.methods[0].kind, "basic");