+++ packages/typespec-azure-portal-core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog - @azure-tools/typespec-azure-portal-core
 
+## 0.56.0
+
+No changes, version bump only.
+
+## 0.55.0
+
+No changes, version bump only.
+
 ## 0.54.0
 
 ### Bump dependencies