+++ packages/typespec-azure-resource-manager/CHANGELOG.md
@@ -1,5 +1,25 @@
 # Change Log - @azure-tools/typespec-azure-resource-manager
 
+## 0.56.0
+
+### Bug Fixes
+
+- [#2250](https://github.com/Azure/typespec-azure/pull/2250) Addressing common type differences with common types in azure-rest-api-specs, such as ordering of elements, lower/upper case in definition names, typos, and incorrect format on the next link.
+- [#2538](https://github.com/Azure/typespec-azure/pull/2538) Fixing gaps in the `@identifiers` decorator functionality:
+  - The `@identifier` decorator should take priority when present, and its value should be respected.
+  - The value of the `@identifier` decorator is determined by the `ModelProperty`, not the array type.
+  - The `@armProviderNamespace` is correctly identified in both scenarios: when applied to the array type or the model property.
+- [#2625](https://github.com/Azure/typespec-azure/pull/2625) Make Patch templates specify implicitOptionality
+
+
+## 0.55.0
+
+### Bug Fixes
+
+- [#2435](https://github.com/Azure/typespec-azure/pull/2435) Change OperationStatusResult.resourceId to armResourceIdentifier
+- [#2564](https://github.com/Azure/typespec-azure/pull/2564) Handle types without nodes
+
+
 ## 0.54.0
 
 ### Features